+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 * added `postcondition`
 * added __slots__ to `ReprableMixin` and `Multirun`
+* added `read_in_file` and `write_to_file`